+++ SummaryStore/store/src/main/java/com/samsung/sra/datastore/ingest/LongIngestBuffer.java
@@ -58,12 +58,12 @@ class LongIngestBuffer implements IngestBuffer {
 
     // FIXME: must reconstruct on SummaryStore reopen
     private transient LongArray timestamps, values;
-    private final long capacity;
+    private final int capacity;
     private int size = 0;
     private final int id;
     private static AtomicInteger num = new AtomicInteger(0);
 
-    LongIngestBuffer(long capacity) {
+    LongIngestBuffer(int capacity) {
         this.capacity = capacity;
         this.id = num.incrementAndGet();
